WPF UI.
Created the UI which allows the end user to update the Account Id and
Device Id settings.
Updated the Installer to install the new UI into a sub directory of the
LMS folder.
Updated the Installer to use the bundle instead of the feature msi.
Update the deploy tool to allow for local installations, instead of
querying for github. This is handy for development.

#135